update entities add cascade and json ignore properties

+++ src/main/java/com/erjiangao/tutorselectiontool/entity/Elective.java
@@ -17,7 +17,7 @@ public class Elective {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private int id;
-    private float grade;
+    private double grade;
 
     @Column(columnDefinition = "TIMESTAMP DEFAULT CURRENT_TIMESTAMP", updatable = false,
             insertable = false)
